move paginate.js into a seperate utils folder
