+++ src/utils/Paginate.js
@@ -1,6 +1,4 @@
-import _ from "lodash";
-
 export function paginate(items, pageNumber, pageSize) {
   const startIndex = (pageNumber - 1) * pageSize;
-  return _(items).slice(startIndex).take(pageSize).value();
+  return items.slice(startIndex, startIndex + pageSize);
 }
